feat: アプリケーション設定を行う箇所を関数か
